Simplified Gamepad (broke API), moved NeoKey
NeoKey should be in IO since that's what it does. The Gamepad is now more like the NeoKey as it just reports read values, not trying to serve up different devices that get used by other things (e.g got rid of all the DeviceFactory stuff).

The Gamepad isn't fully hardware tested as the initial unit seems to be defective.